codeflix-micro-videos/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Kaique Moreira", email = "kaiquemoreiracloud@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.13.1",
    "django>=4.0.6",
    "django-extensions>=3.2.0",
    "dependency-injector>=4.39.1",
    "pytest-django>=4.5.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=1.6.0",
    "pylint>=2.14.1",
    "pytest>=7.1.2",
    "pytest-cov>=3.0.0",
    "black>=22.6.0",
    "model-bakery>=1.6.0",
]
[tool.pdm.scripts]
# black
black = "black ./src"

# pep8
pep8 = "autopep8 --in-place --recursive ./src"

# pylint
pylint = "pylint ./src"

# pytest
pytest = "pytest --ignore __pypackages__"
pytest_cov = "pytest --ignore __pypackages__ --cov ./src --cov-fail-under 80"
pytest_cov_html = "pytest --ignore __pypackages__ --cov ./src --cov-report html:./__coverage"

# django
manage = "python ./src/django_app/manage.py"
migrate = "pdm run manage migrate"
makemigrations = "pdm run manage makemigrations"
runserver = "pdm run manage runserver 0.0.0.0:8000"
show_urls = "pdm run manage show_urls"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
